+++ Core/src/Material/CMakeLists.txt
@@ -8,7 +8,6 @@ target_sources(
         BinnedSurfaceMaterial.cpp
         BinnedSurfaceMaterialAccumulator.cpp
         GridSurfaceMaterial.cpp
-        GridSurfaceMaterialFactory.cpp
         HomogeneousSurfaceMaterial.cpp
         HomogeneousVolumeMaterial.cpp
         Interactions.cpp
